+++ modules/printers.nix
@@ -1,0 +1,8 @@
+{
+  imports = [ "avahi" ];
+
+  nixosModule = { user, pkgs, ... }: {
+    services.printing.enable = true;
+    services.printing.stateless = true;
+  };
+}